Moved PlayerItemAction to ControllerModels
