+++ GameSystemObjects/ControllerModels/PlayerItemActionModel.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace GameSystemObjects.ControllerModels
+{
+    public enum Action
+    {
+        ENABLE, DISABLE, BUY, SELL, UPGRADE
+    }
+
+    public class PlayerActionItemModel
+    { 
+        public string player { get; set; }
+        public string item { get; set; }
+        public Action action { get; set; }
+    }
+}
